// src/api.ts
export type Id = string;

/**
 * Email
 * @format email
 * @example example@example.com
 */
export type Email = string;

/**
 * * Must be at least 8 characters long
 * Must be at most 200 characters long
 * @example password
 */
export type Password = string;

export interface UserForMe {
  /** ID */
  id: Id;

  /** Email */
  email: Email;
}

export interface LoginRequest {
  /** Email */
  email: Email;

  /**
   * * Must be at least 8 characters long
   * * Must be at most 200 characters long
   *
   */
  password: Password;
}

export interface ConfirmEmailToCreateUserRequest {
  /** Email */
  email: Email;
}

export interface SignUpRequest {
  /**
   * * Must be at least 8 characters long
   * * Must be at most 200 characters long
   *
   */
  password: Password;
}
// ...
    export type RequestParams = {};
// ...
export enum ContentType {
  Json = "application/json",
  FormData = "multipart/form-data",
  UrlEncoded = "application/x-www-form-urlencoded",
}
// ...
export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDataType> {
  hello = {
    /**
     * @description Get Hello World
     *
     * @name GetHelloWorld
     * @summary Get Hello World
     * @request GET:/hello
     * @secure
     * @response `200` `Id` Successful operation
     */
    getHelloWorld: (params: RequestParams = {}) =>
      this.request<Id, any>({
        path: `/hello`,
        method: "GET",
        secure: true,
        format: "json",
        ...params,
      }),
  };
  login = {
    /**
     * @description Login
     *
     * @tags authentication
     * @name Login
     * @summary Login
     * @request POST:/login
     * @secure
     * @response `201` `void` Successful operation
     * @response `400` `any`
     * @response `401` `any`
     */
    login: (data: LoginRequest, params: RequestParams = {}) =>
      this.request<UserForMe, any>({
        path: `/login`,
        method: "POST",
        body: data,
        secure: true,
        type: ContentType.Json,
        ...params,
      }),
  };
  logout = {
    /**
     * @description disable cookie to logout
     *
     * @tags authentication
     * @name Logout
     * @summary Logout
     * @request POST:/logout
     * @secure
     * @response `201` `void` Successful operation
     * @response `401` `any`
     */
    logout: (params: RequestParams = {}) =>
      this.request<void, any>({
        path: `/logout`,
        method: "POST",
        secure: true,
        ...params,
      }),
  };
  users = {
    /**
     * @description Confirm email to Create User * Api sends an email to the user with a link to confirm the email address. * if already registerd, successful response is returned too.
     *
     * @tags users
     * @name ConfirmEmailToCreateUser
     * @summary Confirm email to Create User
     * @request POST:/users
     * @secure
     * @response `201` `void` Always returns success if the format of the emaill address is correct
     * @response `400` `any`
     */
    confirmEmailToCreateUser: (data: ConfirmEmailToCreateUserRequest, params: RequestParams = {}) =>
      this.request<void, any>({
        path: `/users`,
        method: "POST",
        body: data,
        secure: true,
        type: ContentType.Json,
        ...params,
      }),

    /**
     * @description Get my user
     *
     * @tags users
     * @name GetMyUser
     * @summary Get my user
     * @request GET:/users/me
     * @secure
     * @response `200` `UserForMe` Successful operation
     * @response `401` `any`
     */
    getMyUser: (params: RequestParams = {}) =>
      this.request<UserForMe, any>({
        path: `/users/me`,
        method: "GET",
        secure: true,
        format: "json",
        ...params,
      }),

    /**
     * @description Sign up
     *
     * @tags users
     * @name SignUp
     * @summary Sign up
     * @request POST:/users/{userId}
     * @secure
     * @response `201` `UserForMe` successful operation
     * @response `400` `any`
     */
    signUp: (userId: Id, data: SignUpRequest, params: RequestParams = {}) =>
      this.request<UserForMe, any>({
        path: `/users/${userId}`,
        method: "POST",
        body: data,
        secure: true,
        type: ContentType.Json,
        format: "json",
        ...params,
      }),
  };
}